requirements.txt
O365==2.0.26
typer==0.9.0
marshmallow-dataclass==8.5.14
PyYAML==6.0
